Removing test_server.py
